==> P3/example02/03-query_filter-zh.ipynb <==
# 过滤、丢弃空值、重复数据删除
使用 `data_08.csv` 和 `data_18.csv`

# 加载数据集
import pandas as pd
df_08 = pd.read_csv('data_08.csv')
df_18 = pd.read_csv('data_18.csv')

# 查看数据集维度
df_08.shape

# 查看数据集维度
df_18.shape

## 按认证区域过滤

# 过滤满足加州标准的行的数据集
df_08 = df_08.query('cert_region == "CA"')
df_18 = df_18.query('cert_region == "CA"')

# 确定唯一的认证区域是加州
df_08['cert_region'].unique()

# 确定唯一的认证区域是加州
df_18['cert_region'].unique()

# 将认证区域列从两个数据集中丢弃
df_08 = df_08.drop('cert_region',axis='columns')
df_18 = df_18.drop('cert_region',axis='columns')

df_08.shape

df_18.shape

## 丢弃含有缺失值的行

# 查看 2008 年每个特征的缺失值数量
df_08.isnull().sum()

# 查看 2018 年每个特征的缺失值数量
df_18.isnull().sum()

# 丢弃两个数据集中有任何空值的行
df_08 = df_08.dropna()
df_18 = df_18.dropna()

# 检查 2008 年的任何列是否有空值 - 应打印为“假”
df_08.isnull().sum().any()

# 检查 2018 年的任何列是否有空值 - 应打印为“假”
df_18.isnull().sum().any()

## 重复数据删除

# 打印 2008 年和 2018 年数据集的重复数量
df_08.duplicated().sum(), df_18.duplicated().sum()

# 丢弃两个数据集中的重复数据
df_08 = df_08.drop_duplicates()
df_18 = df_18.drop_duplicates()

# 再次打印重复数量，确认重复数据已删除——均应为 0 
df_08.duplicated().sum(), df_18.duplicated().sum()


# 保存进度，以便下一段使用
df_08.to_csv('data_08.csv', index=False)
df_18.to_csv('data_18.csv', index=False)
