==> P3/example02/06-fix_datatypes_mpg_greenhouse-zh.ipynb <==
## 修正 `city_mpg`、`hwy_mpg`、`cmb_mpg` 数据类型
    2008 和 2018：将字符串转换成浮点

# 加载数据集
import pandas as pd
df_08 = pd.read_csv('data_08.csv')
df_18 = pd.read_csv('data_18.csv')

# 将 mpg 列转换成浮点
mpg_columns = ['city_mpg','hwy_mpg','cmb_mpg']
for c in mpg_columns:
    df_18[c] = df_18[c].astype(float)
    df_08[c] = df_08[c].astype(float)

## 修正 `greenhouse_gas_score` 数据类型
    2008：将浮点转换成整数

# 将浮点转换成整数
df_08['greenhouse_gas_score'] = df_08['greenhouse_gas_score'].astype('int64')

## 所有数据类型现在都已修正！最后再检查一次，确认所有更改。

df_08.dtypes

df_18.dtypes

df_08.dtypes == df_18.dtypes

# 将新 CLEAN 数据集保存为新文件！
df_08.to_csv('clean_08.csv', index=False)
df_18.to_csv('clean_18.csv', index=False)
